cli: Always print cell count statistics
We now print the statistics to stderr, to later not break json output to
stdout when we start exposing that option. Also, we capture the
statistics before exporting the target cells, as that this the moment
when the statistics are more useful.

This commit also factors out the file extension parsing code, making the
code cleaner and more flexible. I originally made this change because I
believed that I need this flexibility; later this turned out to not be
immediately necessary. However, this is still a step in the right
direction, so I'm keeping the change.